Renames image element in various places
